Migrated from Oro Bundle to standalone symfony console application Moved code to src/ dirctory Updated composer.json to include appropriate dependencies and moved to PSR-4 Namespace configuration Added in bin/dbtools to configure and run the console application Removed all bundle and Oro specific code Added in resources/definitions.yml to contain platform specific sanitization definitions
