--- Aligent/DBToolsBundle/AligentDBToolsBundle.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Aligent\DBToolsBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-/**
- * Bundle class
- *
- * @category  Aligent
- * @package   DBToolsBundle
- * @author    Adam Hall <adam.hall@aligent.com.au>
- * @copyright 2017 Aligent Consulting.
- * @license   https://opensource.org/licenses/mit MIT License
- * @link      http://www.aligent.com.au/
- **/
-class AligentDBToolsBundle extends Bundle
-{
-}
