--- Aligent/DBToolsBundle/DependencyInjection/AligentDBToolsExtension.php
@@ -1,37 +1,0 @@
-<?php
-
-namespace Aligent\DBToolsBundle\DependencyInjection;
-
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-use Symfony\Component\DependencyInjection\Loader;
-/**
- * This is the class that loads and manages your bundle configuration.
- *
- * @category  Aligent
- * @package   DBToolsBundle
- * @author    Adam Hall <adam.hall@aligent.com.au>
- * @copyright 2017 Aligent Consulting.
- * @license   https://opensource.org/licenses/mit MIT License
- * @link      http://www.aligent.com.au/
- * @link http://symfony.com/doc/current/cookbook/bundles/extension.html
- **/
-class AligentDBToolsExtension extends Extension
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function load(array $configs, ContainerBuilder $container)
-    {
-        $configuration = new Configuration();
-        $config = $this->processConfiguration($configuration, $configs);
-        $container->prependExtensionConfig($this->getAlias(), $config);
-
-        $container->setParameter('aligent_db_tools.table_definitions', $config[Configuration::DEFINITIONS]);
-
-        $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('services.yml');
-        $loader->load('commands.yml');
-    }
-}
