--- Aligent/DBToolsBundle/Resources/config/commands.yml
@@ -1,59 +1,0 @@
-services:
-  Aligent\DBToolsBundle\Command\ConsoleCommand:
-    class: Aligent\DBToolsBundle\Command\ConsoleCommand
-    public: false
-    arguments:
-      - '@Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider'
-    tags:
-      - { name: console.command }
-
-  Aligent\DBToolsBundle\Command\CreateCommand:
-    class: Aligent\DBToolsBundle\Command\CreateCommand
-    public: false
-    arguments:
-      - '@Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider'
-    tags:
-      - { name: console.command }
-
-  Aligent\DBToolsBundle\Command\DropCommand:
-    class: Aligent\DBToolsBundle\Command\DropCommand
-    public: false
-    arguments:
-      - '@Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider'
-    tags:
-      - { name: console.command }
-
-  Aligent\DBToolsBundle\Command\DumpCommand:
-    class: Aligent\DBToolsBundle\Command\DumpCommand
-    arguments:
-      - '@Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider'
-      - '@Aligent\DBToolsBundle\Provider\CompressionServiceProvider'
-      - '%aligent_db_tools.definitions%'
-    public: false
-    tags:
-      - { name: console.command }
-
-  Aligent\DBToolsBundle\Command\ImportCommand:
-    class: Aligent\DBToolsBundle\Command\ImportCommand
-    arguments:
-      - '@Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider'
-      - '@Aligent\DBToolsBundle\Provider\CompressionServiceProvider'
-    public: false
-    tags:
-      - { name: console.command }
-
-  Aligent\DBToolsBundle\Command\InfoCommand:
-    class: Aligent\DBToolsBundle\Command\InfoCommand
-    arguments:
-      - '@Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider'
-    public: false
-    tags:
-      - { name: console.command }
-
-  Aligent\DBToolsBundle\Command\QueryCommand:
-    class: Aligent\DBToolsBundle\Command\QueryCommand
-    arguments:
-      - '@Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider'
-    public: false
-    tags:
-      - { name: console.command }
