--- Aligent/DBToolsBundle/Resources/config/oro/bundles.yml
@@ -1,2 +1,0 @@
-bundles:
-    - Aligent\DBToolsBundle\AligentDBToolsBundle
