--- Aligent/DBToolsBundle/Resources/config/services.yml
@@ -1,25 +1,0 @@
-services:
-  Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider:
-    class: Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider
-    arguments:
-      - '%database_name%'
-      - '%database_user%'
-      - '%database_password%'
-      - '%database_host%'
-      - '%database_port%'
-      - '%database_driver%'
-
-  Aligent\DBToolsBundle\Provider\CompressionServiceProvider:
-    class: Aligent\DBToolsBundle\Provider\CompressionServiceProvider
-    calls:
-      - ['addCompressor', ['gz', '@Aligent\DBToolsBundle\Helper\Compressor\Gzip']]
-      - ['addCompressor', ['gzip', '@Aligent\DBToolsBundle\Helper\Compressor\Gzip']]
-      - ['addCompressor', ['none', '@Aligent\DBToolsBundle\Helper\Compressor\Uncompressed']]
-
-  Aligent\DBToolsBundle\Helper\Compressor\Gzip:
-    class: 'Aligent\DBToolsBundle\Helper\Compressor\Gzip'
-    public: true
-
-  Aligent\DBToolsBundle\Helper\Compressor\Uncompressed:
-    class: 'Aligent\DBToolsBundle\Helper\Compressor\Uncompressed'
-    public: true
