+++ src/Command/InfoCommand.php
@@ -3,13 +3,11 @@ namespace Aligent\DBToolsBundle\Command;
 
 use Aligent\DBToolsBundle\Provider\DatabaseConnectionProvider;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\VarDumper\VarDumper;
-
 /**
- * Create Command - Creates an empty database.
+ * Info Command - Displays info on the database connection Oro is using.
  *
  * @category  Aligent
  * @package   DBToolsBundle
@@ -18,11 +16,10 @@ use Symfony\Component\VarDumper\VarDumper;
  * @license   https://opensource.org/licenses/mit MIT License
  * @link      http://www.aligent.com.au/
  **/
-
-class CreateCommand extends Command
+class InfoCommand extends Command
 {
-    const COMMAND_NAME = 'oro:db:create';
-    const COMMAND_DESCRIPTION=  'Creates an empty database.';
+    const COMMAND_NAME = 'info';
+    const COMMAND_DESCRIPTION=  'Dumps the database connection details.';
 
     /**
      * @var DatabaseConnectionProvider
@@ -43,17 +40,10 @@ class CreateCommand extends Command
     /**
      * Configures the name, arguments and options of the command
      */
-    public function configure() {
-        $this
-            ->setName(self::COMMAND_NAME)
-            ->setDescription(self::COMMAND_DESCRIPTION)
-            ->addOption(
-                'only-command',
-                null,
-                InputOption::VALUE_NONE,
-                'Prints only the command. Does not Execute.'
-            )
-        ;
+    protected function configure()
+    {
+        $this->setName(self::COMMAND_NAME)
+            ->setDescription(self::COMMAND_DESCRIPTION);
     }
 
     /**
@@ -61,17 +51,23 @@ class CreateCommand extends Command
      * @param OutputInterface $output
      * @return int|void
      */
-    public function execute(InputInterface $input, OutputInterface $output) {
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
         $connection = $this->connectionProvider->getConnection();
+        $data = [
+            ['host', $connection->getHost()],
+            ['port', $connection->getPort()],
+            ['dbname', $connection->getName()],
+            ['username', $connection->getUser()],
+            ['password', $connection->getPassword()],
+            ['PDO-Connection-String', $connection->getPdoConnectionString()],
+            ['JBDC-Connection-String', $connection->getJbdcConnectionString()],
+            ['cli-String', $connection->getConnectionString()]
+        ];
 
-        $outputOnly = (bool) $input->getOption('only-command');
-        $query = $connection->getCreateDatabaseQuery();
-        $output->writeln($query);
-
-        if (!$outputOnly) {
-            $pdo = $connection->getPDOConnection();
-            $pdo->query($query);
-            $output->writeln('<info>Created database</info> <comment>' . $connection->getName() . '</comment>');
-        }
+        $table = new Table($output);
+        $table->setHeaders(['Name', 'Value'])
+            ->setRows($data);
+        $table->render();
     }
 }
