+++ src/Helper/VerifyOrDie.php
